src/requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-gis
django-cors-headers
django-filter
django-rest-auth
python-decouple
Pillow
django-filter
djangorestframework-simplejwt